Fix warnings

+++ src/CommonDialogs/MultiChoiceDialog/MultiChoiceDialog.cpp
@@ -9,7 +9,7 @@ namespace Examples {
       button->Bind(wxEVT_BUTTON, [&](wxCommandEvent& event) {
         label->SetLabel("");
         std::vector<wxString> choices = {"item 1", "item 2", "item 3", "item 4", "item 5", "item 6", "item 7", "item 8", "item 9", "item 10"};
-        wxMultiChoiceDialog dialog(this, "Message text", "Caption text", choices.size(), choices.data());
+        wxMultiChoiceDialog dialog(this, "Message text", "Caption text", static_cast<int>(choices.size()), choices.data());
         if (dialog.ShowModal() == wxID_OK) {
           for (auto selection : dialog.GetSelections()) {
             if (!label->GetLabel().IsEmpty()) label->SetLabel(wxString::Format("%s\n", label->GetLabel()));
