+++ src/CommonDialogs/SingleChoiceDialog/SingleChoiceDialog.cpp
@@ -9,7 +9,7 @@ public:
     button->Bind(wxEVT_BUTTON, [&](wxCommandEvent& event) {
       label->SetLabel("");
       std::vector<wxString> choices = {"item 1", "item 2", "item 3", "item 4", "item 5", "item 6", "item 7", "item 8", "item 9", "item 10"};
-      wxSingleChoiceDialog dialog(this, "Message text", "Caption text", choices.size(), choices.data());
+      wxSingleChoiceDialog dialog(this, "Message text", "Caption text", static_cast<int>(choices.size()), choices.data());
       if (dialog.ShowModal() == wxID_OK)
         label->SetLabel(dialog.GetStringSelection());
       else
